requirements.txt
asyncpg==0.27.0
fastapi==0.88.0
mangum==0.17.0
pydantic==1.10.4
python-dotenv==1.0.0
python-rapidjson==1.9
requests==2.28.1
tortoise-orm==0.19.2
uvicorn[standard]==0.20.0